move components to separate dir
