+++ src/app/components/admin-page/admin-page.component.ts
@@ -1,8 +1,8 @@
 import { HttpClient } from '@angular/common/http';
 import { Component } from '@angular/core';
 import { finalize } from 'rxjs';
-import { User } from '../models/user';
-import { environment as env } from '../../environments/environment';
+import { User } from '../../models/user';
+import { environment as env } from '../../../environments/environment';
 
 @Component({
   selector: 'app-admin-page',
