+++ src/app/components/dashboard-page/dashboard-page.component.ts
@@ -1,8 +1,8 @@
 import { Component } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { finalize } from 'rxjs';
-import { environment as env } from '../../environments/environment';
-import { Assessment } from '../models/assessment';
+import { environment as env } from '../../../environments/environment';
+import { Assessment } from '../../models/assessment';
 
 @Component({
   selector: 'app-dashboard-page',
