+++ src/app/components/graph-page/graph-page.component.html
