+++ src/app/components/graph-page/graph-page.component.ts
@@ -1,9 +1,9 @@
 import { HttpClient, HttpParams } from '@angular/common/http';
 import { Component } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { environment as env } from '../../environments/environment';
+import { environment as env } from '../../../environments/environment';
 import { finalize } from 'rxjs';
-import { Graph } from '../models/graph';
+import { Graph } from '../../models/graph';
 
 @Component({
   selector: 'app-graph-page',
