+++ src/app/components/loading/loading.component.css
