+++ src/app/components/login-page/login-page.component.ts
@@ -1,10 +1,10 @@
 import { Component, inject } from '@angular/core';
-import { environment as env } from '../../environments/environment';
+import { environment as env } from '../../../environments/environment';
 import { FormBuilder, Validators } from '@angular/forms';
-import { LoginRequest } from '../models/login-request';
+import { LoginRequest } from '../../models/login-request';
 import { finalize } from 'rxjs';
 import { Router } from '@angular/router';
-import { AuthService } from '../services/auth.service';
+import { AuthService } from '../../services/auth.service';
 
 @Component({
   selector: 'app-login-page',
